refactor: rename CompetingProposal to ProposalCard
