+++ src/components/Proposal/View/TenderProposals.tsx
@@ -2,8 +2,7 @@ import React from 'react'
 
 import { ProposalAttributes } from '../../../entities/Proposal/types'
 
-import CompetingProposal from './Budget/CompetingProposal'
-
+import ProposalCard from './ProposalCard'
 import './ProposalProcess.css'
 import Section from './Section'
 
@@ -15,8 +14,7 @@ export default function TenderProposals({ proposals }: Props) {
   return (
     <Section title="Escalated Tender Proposals" isNew>
       {proposals.map((proposal) => {
-        // TODO: Refactor CompetingProposal to ProposalCard inside Proposal/View folder
-        return <CompetingProposal key={proposal.id} proposal={proposal} />
+        return <ProposalCard key={proposal.id} proposal={proposal} />
       })}
     </Section>
   )
